Moved the database cleanup logic into the new database framework.
- Moved the logic in KeepTablesOnly.java into a new method
  (cleanupDatabase()) of the FactorBaseDataBase interface.
- Removed unnecessary code in KeepTablesOnly.java and it now only
  exposes the method that's meant to be called by code outside the
  file (improves encapsulation).
- Removed the try/catch statements in KeepTablesOnly.java so that
  they could be caught and handled more appropriately at a higher
  level in the call stack.
- Updated the query for retrieving the longest RChain.
- A copy of KeepTablesOnly.java was created in the "refactored"
  subdirectory of the obsolete directory for quick reference purposes.
- Updated RunBB.java to use the new cleanup method of the
  FactorBaseDataBase interface.
