+++ code/factorbase/src/main/java/ca/sfu/cs/factorbase/database/FactorBaseDataBase.java
@@ -13,4 +13,13 @@ public interface FactorBaseDataBase {
      * @throws DataBaseException if an error occurs when attempting to access the database.
      */
     void setupDatabase() throws DataBaseException;
+
+
+    /**
+     * This method should cleanup any content (e.g. databases/tables) that are generated by the other
+     * methods of this interface and are no longer needed after FactorBase has finished running.
+     *
+     * @throws DataBaseException if an error occurs when attempting to access the database.
+     */
+    void cleanupDatabase() throws DataBaseException;
 }
